+++ max-bot/src/keyboards.ts
@@ -1,10 +1,9 @@
 import { Keyboard } from "@maxhub/max-bot-api";
-import { env } from "bun";
 
 const PROD_BOT_ID = "id3200014662";
 const DEV_BOT_ID = "id3200014662_2";
 
-const BOT_ID = env.ENV === "development" ? DEV_BOT_ID : PROD_BOT_ID;
+const BOT_ID = process.env.ENV === "development" ? DEV_BOT_ID : PROD_BOT_ID;
 
 export const getMainKeyboard = () => {
   return Keyboard.inlineKeyboard([